android app

+++ service-interface/src/app/document-dialog/document-dialog.component.ts
@@ -2,8 +2,6 @@ import {Component, Inject} from '@angular/core';
 import {MAT_DIALOG_DATA} from "@angular/material/dialog";
 import {DocumentService} from "../services/document.service";
 import {InterventionSheetListComponent} from "../intervention-sheet-list/intervention-sheet-list.component";
-// @ts-ignore
-import { saveAs } from 'file-saver';
 import {DomSanitizer, SafeResourceUrl} from "@angular/platform-browser";
 
 @Component({
@@ -29,17 +27,6 @@ export class DocumentDialogComponent {
     this.loadDocuments();
   }
 
-  // loadDocuments() {
-  //   this.documentService.getDocuments(this.data.intervention.id).subscribe({
-  //     next: (docs) => {
-  //       this.documents = docs.map((doc: any) => ({
-  //         ...doc,
-  //         url: `http://188.24.7.49:8080/api/interventions/${this.data.intervention.id}/documents/${doc.name}`
-  //       }));
-  //     },
-  //     error: (err) => console.error('Error loading documents', err)
-  //   });
-  // }
   loadDocuments(): void {
     this.documentService.getDocuments(this.data.intervention.id).subscribe({
       next: (docs) => {
@@ -52,17 +39,6 @@ export class DocumentDialogComponent {
     });
   }
 
-  // onFileSelected(event: any) {
-  //   const file: File = event.target.files[0];
-  //   if (file) {
-  //     this.documentService.uploadDocument(this.data.intervention.id, file).subscribe({
-  //       next: () => {
-  //         this.loadDocuments(); // reîncarcă documentele după upload
-  //       },
-  //       error: (err) => console.error('Error uploading file', err)
-  //     });
-  //   }
-  // }
   onFileSelected(event: any): void {
     const file: File = event.target.files[0];
     if (file) {
@@ -73,63 +49,35 @@ export class DocumentDialogComponent {
     }
   }
 
-  // deleteDocument(filename: string) {
-  //   this.documentService.deleteDocument(this.data.intervention.id, filename).subscribe({
-  //     next: () => this.loadDocuments(),
-  //     error: (err) => console.error('Delete failed', err)
-  //   });
-  // }
   deleteDocument(filename: string): void {
-    this.documentService.deleteDocument(this.data.intervention.id, filename).subscribe({
-      next: () => this.loadDocuments(),
-      error: (err) => console.error('Delete failed', err)
-    });
+    if(confirm("Sure you want delete it?")) {
+      this.documentService.deleteDocument(this.data.intervention.id, filename).subscribe({
+        next: () => this.loadDocuments(),
+        error: (err) => console.error('Delete failed', err)
+      });
+    }
   }
 
-  // previewDocument(doc: { name: string, url: string }) {
-  //   if (doc.name.toLowerCase().endsWith('.pdf')) {
-  //     this.documentService.getDocumentFile(this.data.intervention.id, doc.name).subscribe({
-  //       next: (blob) => {
-  //         const blobUrl = URL.createObjectURL(blob);
-  //         this.selectedDocumentUrl = blobUrl;
-  //       },
-  //       error: (err) => {
-  //         console.error('Error loading document', err);
-  //         this.selectedDocumentUrl = null;
-  //       }
-  //     });
-  //   } else {
-  //     this.selectedDocumentUrl = null;
-  //     window.open(doc.url, '_blank');
-  //   }
+  // previewDocument(doc: any): void {
+  //   this.loadingPreview = true;
+  //   const url = `http://188.24.7.49:8080/api/interventions/${this.data.intervention.id}/documents/${doc.name}`;
+  //   this.previewUrl = this.sanitizer.bypassSecurityTrustResourceUrl(url);
+  //   this.loadingPreview = false;
   // }
   previewDocument(doc: any): void {
-    this.loadingPreview = true;
     const url = `http://188.24.7.49:8080/api/interventions/${this.data.intervention.id}/documents/${doc.name}`;
-    this.previewUrl = this.sanitizer.bypassSecurityTrustResourceUrl(url);
-    this.loadingPreview = false;
+    window.open(url, '_blank');
   }
 
 
 
-  // downloadDocument(doc: { name: string }) {
-  //   this.documentService.downloadDocument(this.data.intervention.id, doc.name).subscribe({
-  //     next: (blob) => {
-  //       saveAs(blob, doc.name);
-  //     },
-  //     error: (err) => {
-  //       console.error('Download failed', err);
-  //       alert('Descărcarea a eșuat.');
-  //     }
-  //   });
-  // }
-  downloadDocument(document: any) {
-    this.documentService.downloadDocument(this.data.intervention.id, document.name).subscribe({
+  downloadDocument(doc: any) {
+    this.documentService.downloadDocument(this.data.intervention.id, doc.name).subscribe({
       next: (blob) => {
         const url = window.URL.createObjectURL(blob);
         const a = document.createElement('a');
         a.href = url;
-        a.download = document.name;
+        a.download = doc.name;
         a.click();
         window.URL.revokeObjectURL(url);
       },
